+++ src/main/java/project/hospital/domain/PatientRepository.java
@@ -2,6 +2,6 @@ package project.hospital.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PatientRepository extends CrudRepository<Patient, Integer> {
+public interface PatientRepository extends CrudRepository<Patient, Long> {
 	
 }
